rewactoring

+++ ipcounter.go
@@ -11,8 +11,8 @@ const (
 
 type IPCounter interface {
 	Count() uint64
-	Add(ipAddr uint32)
-	AddConcurrent(ipAddr *[]uint32, workerId int)
+	Add(ipAddr uint64)
+	AddConcurrent(ipAddr *[]uint64, workerId int)
 }
 
 type IPCounterBitMap struct {
@@ -25,13 +25,13 @@ func NewBitmapIPCounter() *IPCounterBitMap {
 	}
 }
 
-func (ipc *IPCounterBitMap) Add(ipAddrRepresentation uint32) {
+func (ipc *IPCounterBitMap) Add(ipAddrRepresentation uint64) {
 	chunkNumber := ipAddrRepresentation / bitesIn64Int
 	positionInChunk := ipAddrRepresentation % bitesIn64Int
 	ipc.bitmap[chunkNumber] |= 1 << positionInChunk
 }
 
-func (ipc *IPCounterBitMap) AddConcurrent(ipAddrRepresentation *[]uint32, workerId int) {
+func (ipc *IPCounterBitMap) AddConcurrent(ipAddrRepresentation *[]uint64, workerId int) {
 	// empty
 }
